_Look_Here_/experimental/waitaminute/archive/Old_Unorganized/backup_requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
bandit==1.7.5
bidict==0.22.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dill==0.3.7
dparse==0.6.3
Flask==3.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
netifaces==0.11.0
openai==0.28.1
packaging==23.2
pbr==5.11.1
pipdeptree==2.13.0
platformdirs==3.11.0
Pygments==2.16.1
pylint==3.0.1
pyparsing==3.1.1
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safety==2.3.5
simple-websocket==1.0.0
smmap==5.0.1
stevedore==5.1.0
tomlkit==0.12.1
tqdm==4.66.1
urllib3==2.0.6
Werkzeug==3.0.0
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
